rename file auth.service.ts to auth-service.ts to uniformize
