+++ configuration-client/src/services/http-interceptor.tsx
@@ -1,6 +1,6 @@
 import axios from 'axios';
 import { showNotification } from '@mantine/notifications';
-import { getJwt, redirectToLogin } from './auth.service';
+import { getJwt, redirectToLogin } from './auth-service';
 
 axios.interceptors.request.use(function (request) {
   let jwt = getJwt();
